pyproject.toml
[tool.poetry]
name = "Lightning.py"
version = "4.6.0"
description = ""
authors = ["LightSage <lightsage01@gmail.com>"]
license = "AGPL-3.0-or-later"
readme = "README.md"
repository = "https://gitlab.com/lightning-bot/Lightning"
packages = [
    { include = "lightning/**/*.py" }, 
    { include = "LICENSE" },
    { include = "LICENSE.header" }
]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.12"
"discord.py" = {git = "https://github.com/Rapptz/discord.py.git", branch="master", extras = ["speed"]}
python-dateutil = "^2.8.1"
parsedatetime = "^2.6"
psutil = "^5.7.2"
tabulate = "^0.9.0"
feedparser = "^6.0.10"
asyncpg = "^0.27.0"
discord-ext-menus = {git = "https://github.com/LightSage/discord-ext-menus.git"}
lru_dict = "^1.1.6"
sentry-sdk = ">=1.0.0,<2.0.0"
py-flags = "^1.1.4"
jishaku = "^2.5.0"
rapidfuzz = "^3.0.0"
typer = "^0.9.0"
tomlkit = "^0.11.6"
uvloop = {version = "^0.17.0", platform = "linux"}
orjson = "^3.6.0"
beautifulsoup4 = "^4.9.3"
lxml = "^4.9.2"
sanctum-tc = "^0.7.0"
objgraph = "^3.5.0"
asyncpg-stubs = "^0.27.0"
Wand = "^0.6.11"
rich = "^13.3.2"
setuptools = "^67.6.0"
wheel = "^0.40.0"
matplotlib = "^3.7.0"
numpy = "^1.24.2"
redis = {extras = ["hiredis"], version = "^4.5.4"}
unidecode = "^1.3.6"
prometheus-async = "^22.2.0"

[tool.poetry.dev-dependencies]
isort = "^5.12.0"
flake8 = "^6.0.0"
pre-commit = "^3.2.0"

[tool.poetry.scripts]
lightning = 'lightning.cli.main:parser'

[tool.autopep8]
max_line_length = 120
ignore = ["E301", "F402"]
in-place = true
recursive = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
